+++ zcash_client_backend/src/fees/zip317.rs
@@ -8,12 +8,9 @@ use zcash_primitives::{
     consensus::{self, BlockHeight},
     memo::MemoBytes,
     transaction::{
-        components::{
-            amount::{BalanceError, NonNegativeAmount},
-            sapling::fees as sapling,
-            transparent::fees as transparent,
-        },
+        components::amount::{BalanceError, NonNegativeAmount},
         fees::{
+            transparent,
             zip317::{FeeError as Zip317FeeError, FeeRule as Zip317FeeRule},
             FeeRule,
         },
@@ -21,7 +18,8 @@ use zcash_primitives::{
 };
 
 use super::{
-    ChangeError, ChangeStrategy, ChangeValue, DustAction, DustOutputPolicy, TransactionBalance,
+    sapling, ChangeError, ChangeStrategy, ChangeValue, DustAction, DustOutputPolicy,
+    TransactionBalance,
 };
 
 /// A change strategy that and proposes change as a single output to the most current supported